This PR is consist of two parts:

1. DeepSpeed has 'overlap_comm' option to overlap data transfer and computation during the backward step, but this does not actually work as intended as there is an hidden cuda synchronization. To be more specific, in the `_has_inf_or_nan()` function in `stage2.py` file, the operations related to the variable `cpu_sum` require a 4-byte DtoH data transfer and this is a blocking call. To solve this, I created a pinned memory storage at the initialization step and use this storage as an asynchronous cudaMemcpy target.

2. During the forward pass, partial activations are copied to the host memory through blocking cudaMemcpy calls. Potentially this can be improved with pinned memory allocations, but anyway currently not implemented in such way. Assuming there is a good enough amount of reason to not use the pinned memory allocations for this step, we can at least try to populate the pages of the target tensors ahead of the actual cudaMemcpy call. I have added a simple code which 'touches' the unpopulated pages of the tensor. This approach makes the page population to happen asynchronously to the forward computation time. Without the pre-population step like this, the pages need to be populated during the blocking cudaMemcpy calls, which are already in the critical path of the forward step. The modification is added in `checkpointing.py` file.

In a simple GPT-2 training benchmark, I see the following improvements,
Unmodified: ~20 seconds per iteration
Unmodified + overlap_comm: ~20 seconds per iteration
`_async_has_inf_or_nan()` change + overlap_comm: ~18.5 seconds per iteration
`_async_has_inf_or_nan()` change + overlap_comm + pre-page-population: ~16 seconds per iteration

I think the performance improvement by the second modification can vary significantly depending on the system used to test.