@shhssdm 2020-11-07
> Thanks a ton for putting this together, @shhssdm! Our testing infrastructure needs a bit of work in order to test this confidently. I've been working on this for a bit already, but need a bit of time to prepare some tests before merging this.