@shhssdm 2020-11-07
> Thanks a ton for putting this together, @shhssdm! Our testing infrastructure needs a bit of work in order to test this confidently. I've been working on this for a bit already, but need a bit of time to prepare some tests before merging this.

Sorry while I was trying to reply you I mis-clicked and closed the request but reopened it now. Yeah I really have no problem with it taking longer time for this merge request. I initially thought it would be a very simple request but as @samyam pointed out it required some more structural changes.